fix: Reduce calls on load
