+++ src/state/masterApeMigration/hooks.ts
@@ -57,17 +57,17 @@ export const useSetInitialMigrateStatus = () => {
   const { mergedMigrationLoaded } = useIsMigrationLoading()
   const migrateMaximizers = useMigrateMaximizer()
   // When a user switches accounts we need to reset the initial state by timing the loads to trigger new data
-  useEffect(() => {
-    if (mergedMigrationLoaded) {
-      delay(() => {
-        dispatch(setMigrationLoading({ allDataLoaded: false, userDataLoaded: false, mergedMigrationLoaded: false }))
-      }, 1000)
-      delay(() => {
-        dispatch(setMigrationLoading({ userDataLoaded: true }))
-      }, 10000)
-    }
-    /* eslint-disable react-hooks/exhaustive-deps */
-  }, [dispatch, account])
+  // useEffect(() => {
+  //   if (mergedMigrationLoaded) {
+  //     delay(() => {
+  //       dispatch(setMigrationLoading({ allDataLoaded: false, userDataLoaded: false, mergedMigrationLoaded: false }))
+  //     }, 1000)
+  //     delay(() => {
+  //       dispatch(setMigrationLoading({ userDataLoaded: true }))
+  //     }, 10000)
+  //   }
+  //   /* eslint-disable react-hooks/exhaustive-deps */
+  // }, [dispatch, account])
 
   // Set initial migration status
   useEffect(() => {
